+++ desc-tests/date_tests.ml
@@ -5,49 +5,49 @@ module Alco = struct
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.Ymd.make_exn ~year:1977 ~month:12 ~day:31)
-      (Timedesc.Date.ISO_week.make_exn ~year:1977 ~week:52 ~weekday:`Sat)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1977 ~week:52 ~weekday:`Sat)
 
   let week_date1 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.Ymd.make_exn ~year:1978 ~month:01 ~day:02)
-      (Timedesc.Date.ISO_week.make_exn ~year:1978 ~week:1 ~weekday:`Mon)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1978 ~week:1 ~weekday:`Mon)
 
   let week_date2 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.Ymd.make_exn ~year:1979 ~month:01 ~day:01)
-      (Timedesc.Date.ISO_week.make_exn ~year:1979 ~week:1 ~weekday:`Mon)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1979 ~week:1 ~weekday:`Mon)
 
   let week_date3 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.Ymd.make_exn ~year:1979 ~month:12 ~day:31)
-      (Timedesc.Date.ISO_week.make_exn ~year:1980 ~week:1 ~weekday:`Mon)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1980 ~week:1 ~weekday:`Mon)
 
   let week_date4 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.Ymd.make_exn ~year:1980 ~month:12 ~day:28)
-      (Timedesc.Date.ISO_week.make_exn ~year:1980 ~week:52 ~weekday:`Sun)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1980 ~week:52 ~weekday:`Sun)
 
   let week_date5 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.Ymd.make_exn ~year:1969 ~month:12 ~day:31)
-      (Timedesc.Date.ISO_week.make_exn ~year:1970 ~week:1 ~weekday:`Wed)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1970 ~week:1 ~weekday:`Wed)
 
   let of_iso8601_case0 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.of_iso8601_exn "1977-W52-6")
-      (Timedesc.Date.ISO_week.make_exn ~year:1977 ~week:52 ~weekday:`Sat)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1977 ~week:52 ~weekday:`Sat)
 
   let of_iso8601_case1 () =
     Alcotest.(check date_testable)
       "same date"
       (Timedesc.Date.of_iso8601_exn "1970-W1-3")
-      (Timedesc.Date.ISO_week.make_exn ~year:1970 ~week:1 ~weekday:`Wed)
+      (Timedesc.Date.ISO_week_date.make_exn ~year:1970 ~week:1 ~weekday:`Wed)
 
   let of_iso8601_case2 () =
     Alcotest.(check date_testable)
@@ -117,11 +117,11 @@ module Qc = struct
       ~name:"view_is_same_as_original_iso_week_date" iso_week_date
       (fun (year', week', weekday') ->
         let d =
-          Timedesc.Date.ISO_week.make_exn ~year:year' ~week:week'
+          Timedesc.Date.ISO_week_date.make_exn ~year:year' ~week:week'
             ~weekday:weekday'
         in
-        let { Timedesc.Date.ISO_week.year; week; weekday } =
-          Timedesc.Date.ISO_week.view d
+        let { Timedesc.Date.ISO_week_date.year; week; weekday } =
+          Timedesc.Date.ISO_week_date.view d
         in
         year = year' && week = week' && weekday = weekday')
 
@@ -148,7 +148,7 @@ module Qc = struct
     QCheck.Test.make ~count:100_000 ~name:"iso_week_date_accessors"
       iso_week_date (fun (year', week', weekday') ->
         let d =
-          Timedesc.Date.ISO_week.make_exn ~year:year' ~week:week'
+          Timedesc.Date.ISO_week_date.make_exn ~year:year' ~week:week'
             ~weekday:weekday'
         in
         let year = Timedesc.Date.iso_year d in